Use GitHub Actions for continuous integration
While we had originally planned on using Travis CI with upload-service,
we later switched to GitHub Actions. That was later in the commit
history, and happened via an exploratory process, so instead of
cherry-picking I will just adapt the current version in upload-service.

Credit to Cristina for doing that exploratory work in upload-service!

Co-authored-by: Cristina Muñoz <hi@xmunoz.com>