Connect to PostgreSQL with TinyPg
Set up the initial database configuration with TinyPg[1]. TinyPg is a
lightweight wrapper around the very popular node-postgres[2], adding the
ability to keep SQL queries in separate .sql files. This should allow us
to have first-class editor support as we work with SQL code, as well as
write our queries to use as much or as little of the features PostgreSQL
offers as we choose.

Add documentation about configuration via environment variables, and add
an initial SQL query to validate that the database connection is
successful and that TinyPg is correctly configured to load SQL from
files.

[1] https://joeandaverde.github.io/tinypg/
[2] https://node-postgres.com/

Issue #2 Connect to a database