Set up db migrations with postgres-migrations
Use the postgres-migrations (npm[1], GitHub[2]) library to manage our
database schema. Install the dependency, set up an initial migration,
and ensure migrations are run at application start.

I like postgres-migrations because I find it to be straightforward and
easy to understand. Creating a migration is as simple as saving a SQL
file in the format `number-description.sql`; there are no special
command line tools to run. It keeps its state in the database, so
running a migration is as simple as running a SQL file and recording
that fact in the `migrations` table. We can rely on our existing TinyPg
dependency (which wraps node-postgres) to manage the connection pool,
rather than separately needing to configure our migrations tool.

The initial migration addresses a small flaw in the original schema that
postgres-migrations creates; it is an unfortunate bug that is difficult
for them to fix in a compatible way. Fortunately for us, it serves as a
good example of how to create a migration, and the fix does not require
any library-level changes. See the comments in the migration file for
more details.

Now that we have migrations, we need to keep them separate from the
queries used by TinyPg. Create subdirectories off of src/db/ and
update TinyPg's configuration to look in the new place.

[1] https://www.npmjs.com/package/postgres-migrations
[2] https://github.com/ThomWright/postgres-migrations

Issue #3 Set up database schema migrations