Count ESLint warnings as error
Our CI system, GitHub Actions, looks at the return value of the commands
it runs, and not its output. By default, the ruleset we've adopted
considers some of its rules as warning-level rather than error-level; in
ESLint, warnings do not cause a nonzero return value.

Update our invocation of ESLint to exit with an error status if there
are any warning-level rule violations[1]. We could change just our CI,
but I don't think it harms anything to change our script to error on
warnings.

Also fix the sole warning we currently have, which I had missed
previously - in part because the CI passed.

[1] https://eslint.org/docs/user-guide/command-line-interface#-max-warnings