Tell git not to change source code line endings
By default, git will automatically convert line endings on Windows, so
that it stores the Unix-standard LF characters in the repository and the
Windows-standard CRLF in the checked out copy.

Our tools do NOT like that.

We have ESLint set up with the Airbnb configuration, which requires Unix
linebreak-style[1]. Running it on Windows causes ESLint to report an
error on literally every line of Typescript code, which is noisy and
unhelpful.

More subtly, our migrations library hashes the contents of the migration
files, and stores the value in the database[2]. Changing whitespace in
the migration files, including the line endings, causes the migration
script to error at startup. While the current behavior is stable, in
that repeatedly running with Windows line endings will work, the hashes
stored in the database will differ from those stored in other
developers' databases, and in dev/staging/production.

The solution to both problems is to tell git not to translate our source
code on Windows. Modern Windows developer tools, including VS Code and
the Jetbrains IDEs, know how to deal with Unix line breaks, so this
workaround is not necessary. Configure our .gitattributes file[3] to
always use Unix line endings in our source code.

[1] https://eslint.org/docs/rules/linebreak-style
[2] https://github.com/ThomWright/postgres-migrations#hash-checks-for-previous-migrations
[3] https://git-scm.com/docs/gitattributes