Remove references to .tsx files
The extension .tsx is used for TypeScript files that contain JSX, a
"syntax extension to JavaScript" for HTML-esque code. Since this is a
back-end service which does not currently render anything into HTML, and
does not have JSX as a dependency, we do not need to include .tsx files
in our filters.