Add and configure Firebase
We will be using Firebase Cloud Messaging[1] to send push notifications
to our mobile apps. In order to do that, we first need to add the
Firebase admin SDK[2].

By default, the SDK requires its credentials to live in a JSON file on
disk, and its instructions assume that. However, doing so would
complicate our deployment, as we're currently using GitHub secrets for
managing that kind of credentials. If we need to stuff JSON into an
environment variable anyway - as our current deployment process requires
- we might as well pipe that variable all the way through to the
application. Unfortunately, this requires some additional one-time setup
for developers to populate the environment variable; this is documented
in .env.template.

Initializing the Firebase admin SDK validates that the credentials are
present and well-formed, but does not check that they are valid. FCM has
the concept of sending a "dry-run" message, which does make a network
call and requires valid credentials, so we can use that to make sure
we're configured correctly. Send a dry-run message as part of startup,
and quit if that fails; likewise, send a dry-run message during health
checks.

[1] https://firebase.google.com/products/cloud-messaging
[2] https://firebase.google.com/docs/admin/setup

Issue #17 Add Firebase SDK