Send messages on notification creation
Send push notifications via Firebase Cloud Messaging to each device a
user has registered when a new notification is created.

While the plan was originally to send a data message[1], after some
testing with the mobile team, it seems such messages are not being
received. However, they believe notification messages can be intercepted
by the application before being displayed by the operating system, so
that the system-facing strings can be replaced with user-facing strings.

Use the send[2] API, which appears to be the current supported API for
sending FCM messages. We receive a unique message ID in response, but
since as far as I can tell there is nothing we can do with that message
ID, silently swallow it.

Use setImmediate[3] to queue up the calls to FCM, as we do not need to
block the API request on talking to FCM. If some or all of the calls to
FCM fail, then log an error to standard out; this behavior is not ideal,
and we should set up Sentry to properly record and expose such errors.

There is one error we can handle: invalid or expired tokens. We do not
(yet) validate the device token upon registration, but even if we did,
tokens can expire. If FCM returns an invalid token error, remove it from
the database, as it will never become valid again.

See also the other errors FCM can return[4], which we are currently not
handling beyond logging. Some probably do not need explicit handling,
like "invalid TTL", as we're not setting that option; others, like the
QUOTA_EXCEEDED category, are errors we should handle when we get to the
scale of encountering such errors, and will require more complexity.

[1] https://firebase.google.com/docs/cloud-messaging/concept-options#notifications_and_data_messages
[2] https://firebase.google.com/docs/reference/admin/node/admin.messaging.Messaging-1#send
[3] https://nodejs.org/api/timers.html#timers_setimmediate_callback_args
[4] https://firebase.google.com/docs/reference/fcm/rest/v1/ErrorCode

Issue #18 Send push notifications