Accept, save, and send notification context
Notifications often require more context than just a destination
user and a type. We don't want to be prescriptive about what that
additional context looks like; instead, we want to allow application(s?)
using the notification-service to be able to include and consume
whatever additional data makes sense.

Accept additional context when creating a notification, save it to the
database, and send it as part of the Firebase messages.

We need to do some validation on the context, as the way we are using
Firebase imposes some constraints: some keys are forbidden, and values
must be strings. Use Joi to make sure that the context satisfies those
constraints.

As the constraints are getting a bit complicated, also add
some unit tests to make sure that we're allowing and forbidding the
correct things. As the tests are all very similar, build a new custom
matcher[2], and declare the types for it[3].

Once the input is validated, save it in the database.  Add a new column
to the notifications table using the JSONB data type[1] to store the
additional context.

Finally, send the context data via Firebase, as additional key-value
pairs in the `data` section of the message.

[1] https://www.postgresql.org/docs/current/datatype-json.html
[2] https://jestjs.io/docs/expect#expectextendmatchers
[3] https://javascript.plainenglish.io/jest-how-to-use-extend-with-typescript-4011582a2217

Issue #29 Include notification context

Co-authored-by: Jason Owen <jasonaowen@opentechstrategies.com>