Upgrade TinyPg to 7.0.0
Use the latest version of TinyPg[1]. Most notably, this includes an
update to its dependencies (and therefore our transitive dependencies),
particularly lodash. I believe the lodash advisory[2] did not affect
tinypg, because tinypg does not use the template function which
contained the vulnerability, but it's good to keep `npm audit` happy.

This release also fixes the issue with TinyPg defaults requiring
additional environment variables for peer authentication, so also update
the .env template to reflect that we no longer need to specify the
username.

[1] https://github.com/dynajoe/tinypg/blob/master/CHANGELOG.md#700---2021-05-20
[2] https://www.npmjs.com/advisories/1673