Explicitly depend on ESLint configurations
The configuration for ESLint we want is a combination of multiple
well-known configurations. Currently, the old version of
eslint-config-airbnb-typescript we're depending on appears to implicitly
depend on eslint-config-airbnb-base and eslint-plugin-import, but a
later version does not, which means upgrading would have caused our
configuration to inadvertently most of its rules.

Explicitly extend the airbnb-base ruleset, and explicitly depend on the
import plugin, so that later upgrades are safe and simple.