Update firebase-admin to version 10.3.0
Currently this repo is on firebase-admin's major version 9; the current
major version of this package is 11. It is good to keep dependencies up
to date, and this commit will resolve two dependabot issues at once. The
reason for not updating all the way to major version 11 is that our
current version of node (version 12) stops being supported at that point.