Bump node versions in Github workflows
Use node-version 16 when building and deploying notification-service (it
was originally left on version 12!).

Test on both versions 16 (the version we are currently targeting) and 18
(the version that is currently actively supported) to match how the
upload-service does testing. The notification-service was also testing
multiple node versions before, so this restores that functionality.

PER-9338: Notification-service to Node 16