Remove obsolete lint rules
The airbnb lint rules are not maintained and won't work with various
updates we need to make in this project (node 22, eslint 9, etc).

This commit removes them; it also removes the incorrect "all" lintset
for typescript which is NOT recommended.

This commit functionally means we are no longer benefitting from
linting, which needs to be resolved as soon as possible.

Issue #121 Update to ESLint 9