Enable auto-merge for Dependabot PRs
We have robust rules on our main branch which requires (1) approval from
a reviewer and (2) CI to pass.  This means that it's safe for auto-merge
to be enabled for dependabot PRs; arguably actually safer than a manual
merge since this workflow will check that the PR was authored by
dependabot and not a spoofer.

Issue #199 Set up dependabot PRs as "auto-merge"